Extract SQL query management from IcebergConnection into QueryManager
